requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
Django==4.2
django-debug-toolbar==4.4.6
exceptiongroup==1.2.2
executing==2.2.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2==2.8.6
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pywin32==308
pyzmq==26.2.1
six==1.17.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
zipp==3.21.0
